requirements.txt
Django==1.11.14
django-admin==1.3.0
django-excel-response2==2.0.8
django-six==1.0.4
djangorestframework==3.8.2
elasticsearch==6.3.0
elasticsearch-dsl==6.2.1